requirements.txt
anthropic>=0.40.0
streamlit>=1.40.0
chromadb>=0.5.0
sentence-transformers>=2.2.0
requests>=2.31.0
beautifulsoup4>=4.12.0
markdownify>=0.13.0
python-dotenv>=1.0.0
lxml>=4.9.0
pdfplumber>=0.10.0
python-docx>=0.8.11
